переделал вывод фильма на класс
